backend/requirements.txt
fastapi
uvicorn
pymongo
psycopg2-binary
openai
google-generativeai
python-dotenv
pydantic
